Refactor: Extract constants, types, and helpers into separate modules
