+++ src/usecases/messages.ts
@@ -1,0 +1,26 @@
+export const MessageTemplates = {
+  tokenDeleteSuccess: "トークン情報を削除しました",
+  tokenDeleteFailure: "トークン情報の削除に失敗しました",
+  postbackAddReply: (calendarName: string) => `『${calendarName}』を購読カレンダーに追加しました。`,
+  postbackAddResult: "カレンダー追加を完了しました",
+  postbackDeleteReply: (name: string) => `『${name}』を購読カレンダーから削除しました。`,
+  postbackDeleteResult: "カレンダー削除を完了しました",
+  calendarListHeader: "購読中のカレンダー:",
+  noSubscribedCalendars: "購読中のカレンダーはありません。『カレンダー追加』で登録できます。",
+  sendAuthGuidance: "Googleカレンダーとの連携を開始します。以下のURLをクリックして認可を行ってください：",
+  sendAuthUrlResult: "認可URLを送信しました",
+  addQuickPrompt: "追加するカレンダーを選択してください",
+  addQuickResult: "カレンダー追加クイックリプライを送信しました",
+  deleteQuickPrompt: "削除するカレンダーを選択してください",
+  deleteQuickResult: "カレンダー削除クイックリプライを送信しました",
+  deleteNoTargetResult: "削除対象のカレンダーがありませんでした",
+  noEvent: "no event",
+  unsupportedMessage: "未対応のメッセージです",
+  postbackError: "ポストバック処理でエラーが発生しました",
+  // Failure messages for unified error handling
+  calendarListFailure: "カレンダー一覧の返信に失敗しました",
+  authUrlSendFailure: "認可URLの送信に失敗しました",
+  addQuickFailure: "カレンダー追加クイックリプライの送信に失敗しました",
+  deleteQuickFailure: "カレンダー削除クイックリプライの送信に失敗しました",
+  tokenFetchFailure: "認可状態の取得に失敗しました",
+} as const;
